Add Flow Based Bilateral filter
Move some common fonction in util file, improve
the fDoG edges by taking round value when moving
between pixel instead of floor value

+++ include/tensor.h
@@ -16,7 +16,7 @@ return, the eigen values need to be given as pointers.
 @param gAmplitude, the gradient amplitude
 @param iAmplitude, the isophote amplitude
 */
-void eigen(const Mat & tensor, Point2d* gradient, Point2d* normal, float* gAmplitude, float* iAmplitude){
+void eigen(const Mat & tensor, Point2f* gradient, Point2f* normal, float* gAmplitude, float* iAmplitude){
   //compute Delta as D=(g11-g22)²+4*(g12)²
   float delta = pow(tensor.at<float>(0,0)-tensor.at<float>(1,1),2)
                 + 4*pow(tensor.at<float>(0,1),2);
